Fix sign-compare warning

+++ test/pi_table.cpp
@@ -57,7 +57,7 @@ int main()
 
   primesieve::iterator it;
   uint64_t prime = it.next_prime();
-  uint64_t count = 1;
+  int count = 1;
 
   while (prime < pi.size())
   {
